// src/generated/services/Krbork_savingsgoalsService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import type { Krbork_savingsgoalsBase, Krbork_savingsgoals } from '../models/Krbork_savingsgoalsModel';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Krbork_savingsgoalsService {
  private static readonly dataSourceName = 'krbork_savingsgoals';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Krbork_savingsgoalsBase, 'krbork_savingsgoalid'>): Promise<IOperationResult<Krbork_savingsgoals>> {
    const result = await Krbork_savingsgoalsService.client.createRecordAsync<Omit<Krbork_savingsgoalsBase, 'krbork_savingsgoalid'>, Krbork_savingsgoals>(
      Krbork_savingsgoalsService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Krbork_savingsgoalsBase, 'krbork_savingsgoalid'>>): Promise<IOperationResult<Krbork_savingsgoals>> {
    const result = await Krbork_savingsgoalsService.client.updateRecordAsync<Partial<Omit<Krbork_savingsgoalsBase, 'krbork_savingsgoalid'>>, Krbork_savingsgoals>(
      Krbork_savingsgoalsService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Krbork_savingsgoalsService.client.deleteRecordAsync(
      Krbork_savingsgoalsService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Krbork_savingsgoals>> {
    const result = await Krbork_savingsgoalsService.client.retrieveRecordAsync<Krbork_savingsgoals>(
      Krbork_savingsgoalsService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Krbork_savingsgoals[]>> {
    const result = await Krbork_savingsgoalsService.client.retrieveMultipleRecordsAsync<Krbork_savingsgoals>(
      Krbork_savingsgoalsService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Krbork_savingsgoals> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Krbork_savingsgoalsService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Krbork_savingsgoalsService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}