// src/generated/services/Krbork_depositsService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import type { Krbork_depositsBase, Krbork_deposits } from '../models/Krbork_depositsModel';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Krbork_depositsService {
  private static readonly dataSourceName = 'krbork_deposits';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Krbork_depositsBase, 'krbork_depositid'>): Promise<IOperationResult<Krbork_deposits>> {
    const result = await Krbork_depositsService.client.createRecordAsync<Omit<Krbork_depositsBase, 'krbork_depositid'>, Krbork_deposits>(
      Krbork_depositsService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Krbork_depositsBase, 'krbork_depositid'>>): Promise<IOperationResult<Krbork_deposits>> {
    const result = await Krbork_depositsService.client.updateRecordAsync<Partial<Omit<Krbork_depositsBase, 'krbork_depositid'>>, Krbork_deposits>(
      Krbork_depositsService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Krbork_depositsService.client.deleteRecordAsync(
      Krbork_depositsService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Krbork_deposits>> {
    const result = await Krbork_depositsService.client.retrieveRecordAsync<Krbork_deposits>(
      Krbork_depositsService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Krbork_deposits[]>> {
    const result = await Krbork_depositsService.client.retrieveMultipleRecordsAsync<Krbork_deposits>(
      Krbork_depositsService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Krbork_deposits> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Krbork_depositsService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Krbork_depositsService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}