// src/generated/models/Krbork_savingsgoalsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Krbork_savingsgoalsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Krbork_savingsgoalsstatecode = keyof typeof Krbork_savingsgoalsstatecode;
export const Krbork_savingsgoalsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Krbork_savingsgoalsstatuscode = keyof typeof Krbork_savingsgoalsstatuscode;

export interface Krbork_savingsgoalsBase {
  importsequencenumber?: number;
  krbork_deadline?: string;
  krbork_name?: string;
  krbork_savingsgoalid: string;
  krbork_targetamount?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Krbork_savingsgoalsstatecode;
  statuscode?: Krbork_savingsgoalsstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Krbork_savingsgoals extends Krbork_savingsgoalsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: number;
  krbork_targetamount_base?: number;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}