// src/generated/models/Krbork_depositsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Krbork_depositsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Krbork_depositsstatecode = keyof typeof Krbork_depositsstatecode;
export const Krbork_depositsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Krbork_depositsstatuscode = keyof typeof Krbork_depositsstatuscode;

export interface Krbork_depositsBase {
  importsequencenumber?: number;
  krbork_amount?: number;
  krbork_depositdate?: string;
  krbork_depositid: string;
  krbork_name?: string;
  "krbork_SavingsGoalId@odata.bind"?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Krbork_depositsstatecode;
  statuscode?: Krbork_depositsstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Krbork_deposits extends Krbork_depositsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: number;
  krbork_amount_base?: number;
  krbork_savingsgoalidname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  krbork_savingsgoalid?: object;
  _krbork_savingsgoalid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}